backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.12.0
qdrant-client==1.7.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
tiktoken==0.5.2
pydantic==2.5.3
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
email-validator==2.1.0
slowapi==0.1.9
